use datatype torch.long for y

+++ sktime/classification/deep_learning/_pytorch.py
@@ -281,5 +281,5 @@ class PytorchDataset(Dataset):
 
         # return y during fit
         y = self.y[i]
-        y = torch.tensor(y)
+        y = torch.tensor(y, dtype=torch.long)
         return x, y
